+++ cmake/modules/DuneAddLibrary.cmake
@@ -20,6 +20,7 @@ Add a library to a Dune module.
       [COMPILE_OPTIONS "<flags>;..."]
       [OUTPUT_NAME <libname>]
       [EXPORT_NAME <exportname>]
+      [NAMESPACE <namespace>]
       [NO_EXPORT]
       [NO_MODULE_LIBRARY]
     )
@@ -28,10 +29,9 @@ Add a library to a Dune module.
     this created ``lib<libname>.so`` or ``lib<libname>.a``. The target properties
     are automatically filled with the given (optional) arguments.
 
-    A dune library is (by default) exported into the ``<export-set>`` given by the
-    global name ``${ProjectName}-targets`` if the parameter ``NO_EXPORT`` is not
-    given. This ``<export-set>`` is automatically installed and exported in the
-    ``dune_finalize_project()`` function.
+    A dune library is exported into a unique export set if the parameter
+    ``NO_EXPORT`` is not given. This export set is automatically installed and
+    exported in the ``dune_finalize_project()`` function.
 
   ``SOURCES``
     The source files from which to build the library.
@@ -45,14 +45,24 @@ Add a library to a Dune module.
   ``OUTPUT_NAME``
     Name of the library file, e.g. ``lib<libname>.so`` or ``lib<libname>.a``.
 
+  ``NAMESPACE``
+    Name to be prepended to the export name of the target.
+    By default this is set to ``Dune::``.
+
   ``EXPORT_NAME``
     Name of the exported target to be used when linking against the library.
+    The name is prepended with the given namespace (e.g. the default Dune::).
+    We recommend to choose an export name with a camel title case matching your
+    library name (e.g., Common, ISTL, and MultiDomainGrid will be exported as
+    Dune::Common, Dune::ISTL, and Dune::MultiDomainGrid)
 
   ``NO_EXPORT``
     If omitted the library is exported for usage in other modules.
 
   ``NO_MODULE_LIBRARY``
-    If omitted the library is added to the global property ``<module>_LIBRARIES``.
+    If omitted the library is added to the global property ``<module>_LIBRARIES``
+    and used for internal module configuration. If ``NO_EXPORT`` is omitted, the library
+    will additionally be added to the global property ``<module>_INTERFACE_LIBRARIES``.
 
 
   .. code-block:: cmake
@@ -61,6 +71,7 @@ Add a library to a Dune module.
       [LINK_LIBRARIES <targets>...]
       [COMPILE_OPTIONS "<flags>;..."]
       [EXPORT_NAME <exportname>]
+      [NAMESPACE <namespace>]
       [NO_EXPORT]
       [NO_MODULE_LIBRARY]
     )
@@ -69,20 +80,34 @@ Add a library to a Dune module.
     An interface target does not contain any sources but my contain flags and
     dependencies.
 
+    A dune library is exported into a unique export set if the parameter
+    ``NO_EXPORT`` is not given. This export set is automatically installed and
+    exported in the ``dune_finalize_project()`` function.
+
   ``LINK_LIBRARIES``
     A list of dependency the libraries is explicitly linked against.
 
   ``COMPILE_OPTIONS``
     Any additional compile flags for building the library.
 
+  ``NAMESPACE``
+    Name to be prepended to the export name of the target.
+    By default this is set to ``Dune::``.
+
   ``EXPORT_NAME``
     Name of the exported target to be used when linking against the library.
+    The name is prepended with the given namespace (e.g. the default Dune::).
+    We recommend to choose an export name with a camel title case matching your
+    library name (e.g., Common, ISTL, and MultiDomainGrid will be exported as
+    Dune::Common, Dune::ISTL, and Dune::MultiDomainGrid)
 
   ``NO_EXPORT``
     If omitted the library is exported for usage in other modules.
 
   ``NO_MODULE_LIBRARY``
-    If omitted the library is added to the global property ``<module>_LIBRARIES``.
+    If omitted the library is added to the global property ``<module>_LIBRARIES``
+    and used for internal module configuration. If ``NO_EXPORT`` is omitted, the library
+    will additionally be added to the global property ``<module>_INTERFACE_LIBRARIES``.
 
 
   .. code-block:: cmake
@@ -134,7 +159,7 @@ endfunction(dune_add_library)
 function(dune_add_library_normal _name)
   cmake_parse_arguments(ARG
     "NO_EXPORT;NO_MODULE_LIBRARY;STATIC;SHARED;MODULE"
-    "COMPILE_FLAGS;COMPILE_OPTIONS;OUTPUT_NAME;EXPORT_NAME"
+    "COMPILE_FLAGS;COMPILE_OPTIONS;OUTPUT_NAME;EXPORT_NAME;NAMESPACE"
     "ADD_LIBS;LINK_LIBRARIES;SOURCES" ${ARGN})
   list(APPEND ARG_SOURCES ${ARG_UNPARSED_ARGUMENTS})
   dune_expand_object_libraries(ARG_SOURCES ARG_ADD_LIBS ARG_COMPILE_FLAGS)
@@ -172,24 +197,42 @@ function(dune_add_library_normal _name)
   # Prepare the export of the library
   if(NOT ARG_NO_EXPORT)
     if(NOT ARG_EXPORT_NAME)
+      message(DEPRECATION
+        "The function dune_add_library(<lib> ...) now requires to provide NO_EXPORT or EXPORT_NAME. "
+        "We recommend to choose an export name with a camel title case matching your library name "
+        "(e.g., Common, ISTL, and MultiDomainGrid will be exported as Dune::Common, Dune::ISTL, and Dune::MultiDomainGrid)\n"
+        " * Calls to `dune_add_library(<lib> ...)` without export specification will be supported until Dune 2.11\n"
+        " * Consumption of unscoped targets `<lib>` will be supported until Dune 2.12")
       set(ARG_EXPORT_NAME ${_name})
     endif()
 
-    set_target_properties(${_name} PROPERTIES EXPORT_NAME ${ARG_EXPORT_NAME})
-
-    # Register library in global property <module>LIBRARIES
-    if(NOT ARG_NO_MODULE_LIBRARY)
-      set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${ARG_EXPORT_NAME})
+    if(NOT ARG_NAMESPACE)
+      set(ARG_NAMESPACE Dune::)
     endif()
 
-    # Install targets to use the libraries in other modules.
-    install(TARGETS ${_name}
-      EXPORT ${ProjectName}-targets DESTINATION ${CMAKE_INSTALL_LIBDIR})
+    set(alias ${ARG_NAMESPACE}${ARG_EXPORT_NAME})
+    if(NOT TARGET ${alias})
+      add_library(${alias} ALIAS ${_name})
+    else()
+      message(WARNING "Target `${_name}` could not be aliased as `${alias}`")
+    endif()
 
-    set(${ProjectName}_EXPORT_SET ${ProjectName}-targets CACHE INTERNAL "")
-  elseif(NOT ARG_NO_MODULE_LIBRARY)
-    # Register library in global property <module>_LIBRARIES
-    set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${_name})
+    set(export_set ${ProjectName}-${ARG_NAMESPACE}-export-set)
+
+    # Install targets to use the libraries in other modules.
+    set_target_properties(${_name} PROPERTIES EXPORT_NAME ${ARG_EXPORT_NAME})
+    install(TARGETS ${_name}
+      EXPORT ${export_set} DESTINATION ${CMAKE_INSTALL_LIBDIR})
+
+    # Register target as an exported library
+    if(NOT ARG_NO_MODULE_LIBRARY)
+      set_property(GLOBAL APPEND PROPERTY ${ProjectName}_INTERFACE_LIBRARIES ${alias})
+    endif()
+  endif()
+
+  # Register library in global property <module>_LIBRARIES
+  if(NOT ARG_NO_MODULE_LIBRARY)
+    set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${name})
   endif()
 endfunction(dune_add_library_normal)
 
@@ -198,7 +241,7 @@ endfunction(dune_add_library_normal)
 function(dune_add_library_interface _name)
   cmake_parse_arguments(ARG
     "NO_EXPORT;NO_MODULE_LIBRARY;INTERFACE"
-    "COMPILE_FLAGS;COMPILE_OPTIONS;EXPORT_NAME"
+    "COMPILE_FLAGS;COMPILE_OPTIONS;EXPORT_NAME;NAMESPACE"
     "ADD_LIBS;LINK_LIBRARIES" ${ARGN})
   list(APPEND ARG_LINK_LIBRARIES ${ARG_ADD_LIBS})
   list(APPEND ARG_COMPILE_OPTIONS ${ARG_COMPILE_FLAGS})
@@ -215,24 +258,42 @@ function(dune_add_library_interface _name)
   # Prepare the export of the library
   if(NOT ARG_NO_EXPORT)
     if(NOT ARG_EXPORT_NAME)
+      message(DEPRECATION
+        "The function dune_add_library(<lib> ...) now requires to provide NO_EXPORT or EXPORT_NAME. "
+        "We recommend to choose an export name with a camel title case matching your library name "
+        "(e.g., Common, ISTL, and MultiDomainGrid will be exported as Dune::Common, Dune::ISTL, and Dune::MultiDomainGrid)\n"
+        " * Calls to `dune_add_library(<lib> ...)` without export specification will be supported until Dune 2.11\n"
+        " * Consumption of unscoped targets `<lib>` will be supported until Dune 2.12")
       set(ARG_EXPORT_NAME ${_name})
     endif()
 
-    set_target_properties(${_name} PROPERTIES EXPORT_NAME ${ARG_EXPORT_NAME})
-
-    # Register library in global property <module>_LIBRARIES
-    if(NOT ARG_NO_MODULE_LIBRARY)
-      set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${ARG_EXPORT_NAME})
+    if(NOT ARG_NAMESPACE)
+      set(ARG_NAMESPACE Dune::)
     endif()
 
-    # Install targets to use the libraries in other modules.
-    install(TARGETS ${_name}
-      EXPORT ${ProjectName}-targets DESTINATION ${CMAKE_INSTALL_LIBDIR})
+    set(alias ${ARG_NAMESPACE}${ARG_EXPORT_NAME})
+    if(NOT TARGET ${alias})
+      add_library(${alias} ALIAS ${_name})
+    else()
+      message(WARNING "Target `${_name}` could not be aliased as `${alias}`")
+    endif()
 
-    set(${ProjectName}_EXPORT_SET ${ProjectName}-targets CACHE INTERNAL "")
-  elseif(NOT ARG_NO_MODULE_LIBRARY)
-    # Register library in global property <module>_LIBRARIES
-    set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${_name})
+    set(export_set ${ProjectName}-${ARG_NAMESPACE}-export-set)
+
+    # Install targets to use the libraries in other modules.
+    set_target_properties(${_name} PROPERTIES EXPORT_NAME ${ARG_EXPORT_NAME})
+    install(TARGETS ${_name}
+      EXPORT ${export_set} DESTINATION ${CMAKE_INSTALL_LIBDIR})
+
+    # Register target as an exported library
+    if(NOT ARG_NO_MODULE_LIBRARY)
+      set_property(GLOBAL APPEND PROPERTY ${ProjectName}_INTERFACE_LIBRARIES ${alias})
+    endif()
+  endif()
+
+  # Register library in global property <module>_LIBRARIES
+  if(NOT ARG_NO_MODULE_LIBRARY)
+    set_property(GLOBAL APPEND PROPERTY ${ProjectName}_LIBRARIES ${name})
   endif()
 endfunction(dune_add_library_interface)